requirements.txt
flask
tensorflow
numpy
pillow
requests
beautifulsoup4
